+++ doc/changelog.rst
@@ -18,6 +18,7 @@ Features
 - Added DNS_ENABLED module into DC settings (API & GUI) - `#45 <https://github.com/erigones/esdc-ce/issues/45>`__
 - Exposed compute node, network and image UUIDs via API - `#49 <https://github.com/erigones/esdc-ce/issues/49>`__
 - Added harvest_vm function into API documentation - `#51 <https://github.com/erigones/esdc-ce/issues/51>`__
+- Made image server optional and configurable (``VMS_IMAGE_VM``) - `#52 <https://github.com/erigones/esdc-ce/issues/52>`__
 
 
 Bugs